+++ CHANGELOG.markdown
@@ -1,3 +1,4 @@
 # 0.2.7 Sun Bed
 
 - Port forwarding from 80 to 8080
+- Create and assign elastic IP address to instances